feat: add Telegram bot Mini App integration

+++ apps/api/alembic/versions/0001_init.py
@@ -17,7 +17,7 @@ depends_on = None
 def upgrade() -> None:
     op.create_table('users',
         sa.Column('id', sa.Integer(), primary_key=True),
-        sa.Column('telegram_id', sa.Integer(), nullable=False),
+        sa.Column('telegram_id', sa.BigInteger(), nullable=False),
         sa.Column('first_name', sa.String(length=255)),
         sa.Column('last_name', sa.String(length=255)),
         sa.Column('username', sa.String(length=255)),
